+++ src/wtforms/__init__.py
@@ -4,4 +4,4 @@ from wtforms.fields import *
 from wtforms.form import Form
 from wtforms.validators import ValidationError
 
-__version__ = "3.0.0.a1"
+__version__ = "3.0.0"